Revisit day 23 part 2
When I first submitted day 23, I was a bit surprised that the algorithm I had got the answer, given I expected it would veer into a wrong answer early and be unrecoverable, but it did manage to get the answer to the problem.

However, there are inputs for which it isn't as lucky and doesn't work. Since then, I've learned of another algorithm, the Bron-Kerbosch algorithm, which handles this quite well and more correctly. I wanted to have a go at fixing my day 23 submission, so I did.